requirements.txt
streamlit
plotly
google-generativeai
langchain-text-splitters
yt-dlp
